fix: adjusted giropay and co-branded bcmc tests
